Time: 0 ms (100%), Space: 40.7 MB (37.3%) - LeetHub

+++ 2579-count-total-number-of-colored-cells/2579-count-total-number-of-colored-cells.java
@@ -1,6 +1,6 @@
 class Solution {
     public long coloredCells(int N) {
         long n = (long)N;
-        return 1 + (long)4 * (n*(n-1)/2);
+        return 1 + 4 * (n*(n-1)/2);
     }
 }
